Waypoints now use new font system

+++ src/main/java/minegame159/meteorclient/rendering/MyFont.java
@@ -110,7 +110,7 @@ public class MyFont {
         return getWidth(string, string.length());
     }
 
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
